Move fidget-jit to a standalone crate
