+++ Cargo.toml
@@ -5,6 +5,8 @@ members = [
     "fidget",
     "fidget-core",
     "fidget-rhai",
+    "fidget-mesh",
+    "fidget-jit",
 
     # Non-WASM demos
     "demos/constraints",
@@ -18,6 +20,12 @@ exclude = ["demos/web-editor/crate"]
 
 [workspace.package]
 rust-version = "1.87"
+version = "0.4.0"
+edition = "2024"
+license = "MPL-2.0"
+repository = "https://github.com/mkeeter/fidget"
+description = "Infrastructure for complex closed-form implicit surfaces"
+authors = ["Matt Keeter <matt.j.keeter@gmail.com"]
 
 [profile.release]
 debug = true
