+++ fidget-core/Cargo.toml
@@ -1,12 +1,13 @@
 [package]
 name = "fidget-core"
-version = "0.4.0"
-edition = "2024"
-license = "MPL-2.0"
-repository = "https://github.com/mkeeter/fidget"
 description = "Infrastructure for complex closed-form implicit surfaces"
-authors = ["Matt Keeter <matt.j.keeter@gmail.com"]
 readme = "../README.md"
+
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+authors.workspace = true
 rust-version.workspace = true
 
 [dependencies]
@@ -25,10 +26,6 @@ zerocopy.workspace = true
 
 workspace-hack = { version = "0.1", path = "../workspace-hack" }
 
-# JIT
-dynasmrt = { workspace = true, optional = true }
-libc = { workspace = true, optional = true }
-
 [target.'cfg(target_os = "windows")'.dependencies]
 windows.workspace = true
 
@@ -39,16 +36,6 @@ getrandom-02 = { package = "getrandom", version = "0.2", features = ["js"] }
 rhai = { workspace = true, optional = true, features = ["wasm-bindgen"] }
 
 [features]
-default = ["jit"]
-
-## Enables fast evaluation via a JIT compiler.  This is exposed in the
-## [`fidget::jit`](crate::jit) module, and is supported on macOS, Linux, and
-## Windows (i.e. all supported platforms except WebAssembly).  There's no way to
-## disable the feature on specific platforms
-## ([Cargo issue](https://github.com/rust-lang/cargo/issues/1197));
-## users will have to disable it manually via `default-features = false`.
-jit = ["dep:dynasmrt", "dep:libc"]
-
 ## Enable `eval-tests` if you're writing your own evaluators and want to
 ## unit-test them.  When enabled, the crate exports a set of macros to test each
 ## evaluator type, e.g. `float_slice_tests!(...)`.
