+++ fidget-core/src/core/types/grad.rs
@@ -238,7 +238,7 @@ impl Grad {
     }
 
     /// Checks that the two values are roughly equal, panicking otherwise
-    #[cfg(test)]
+    #[cfg(any(test, feature = "eval-tests"))]
     pub(crate) fn compare_eq(&self, other: Self) {
         let d = (self.v - other.v)
             .abs()
