+++ fidget-jit/src/aarch64/grad_slice.rs
@@ -1,12 +1,9 @@
 use crate::{
-    Error,
-    jit::{
-        Assembler, AssemblerData, IMM_REG, OFFSET, REGISTER_LIMIT,
-        grad_slice::GradSliceAssembler, mmap::Mmap, reg,
-    },
-    types::Grad,
+    Assembler, AssemblerData, IMM_REG, OFFSET, REGISTER_LIMIT,
+    grad_slice::GradSliceAssembler, mmap::Mmap, reg,
 };
-use dynasmrt::{DynasmApi, DynasmLabelApi, dynasm};
+use dynasmrt::{DynasmApi, DynasmError, DynasmLabelApi, dynasm};
+use fidget_core::types::Grad;
 
 /// Implementation for the gradient slice assembler on `aarch64`
 ///
@@ -509,7 +506,7 @@ impl Assembler for GradSliceAssembler {
         IMM_REG.wrapping_sub(OFFSET)
     }
 
-    fn finalize(mut self) -> Result<Mmap, Error> {
+    fn finalize(mut self) -> Result<Mmap, DynasmError> {
         dynasm!(self.0.ops
             // update our "items remaining" counter
             ; sub x2, x2, 1 // We handle 1 item at a time
