+++ fidget-jit/src/aarch64/point.rs
@@ -1,12 +1,8 @@
 use crate::{
-    Error,
-    jit::{
-        Assembler, AssemblerData, CHOICE_BOTH, CHOICE_LEFT, CHOICE_RIGHT,
-        IMM_REG, OFFSET, REGISTER_LIMIT, mmap::Mmap, point::PointAssembler,
-        reg,
-    },
+    Assembler, AssemblerData, CHOICE_BOTH, CHOICE_LEFT, CHOICE_RIGHT, IMM_REG,
+    OFFSET, REGISTER_LIMIT, mmap::Mmap, point::PointAssembler, reg,
 };
-use dynasmrt::{DynasmApi, dynasm};
+use dynasmrt::{DynasmApi, DynasmError, dynasm};
 
 /// Implementation for the single-point assembler on `aarch64`
 ///
@@ -456,7 +452,7 @@ impl Assembler for PointAssembler {
         IMM_REG.wrapping_sub(OFFSET)
     }
 
-    fn finalize(mut self) -> Result<Mmap, Error> {
+    fn finalize(mut self) -> Result<Mmap, DynasmError> {
         if self.0.saved_callee_regs {
             dynasm!(self.0.ops
                 // Restore callee-saved registers
