+++ fidget-jit/src/interval.rs
@@ -1,3 +1,3 @@
-use crate::jit::AssemblerData;
+use crate::AssemblerData;
 
 pub struct IntervalAssembler(pub(crate) AssemblerData<[f32; 2]>);
