+++ fidget-jit/src/point.rs
@@ -1,3 +1,3 @@
-use crate::jit::AssemblerData;
+use crate::AssemblerData;
 
 pub struct PointAssembler(pub(crate) AssemblerData<f32>);
