+++ fidget-jit/src/x86_64/grad_slice.rs
